still haven't thought of a good way to handle a stroke in progess
it seems an IORef (or some sort of mutable reference) is necessary for
the document as a whole since we need it in the drawCb and we need it to
persist across calls to drawCb

as discovered last night, we probably can't do any cairo drawing outside
of drawCb, which is fine but we need to draw in progess strokes

so we can use an IORef for the document and another IORef for the stroke
in progess? And modify the stroke IORef in the stylusCbs?
The stroke IORef may need to actually be more general to accomodate
rectangular selections eventually. So we might need a larger state
package w/ not only the IORef for the document and stroke but enums for
conditional extentions